fix : 修复代码无法展示

+++ src/main/java/com/music/entity/Admin.java
@@ -1,6 +1,8 @@
 package com.music.entity;
 
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -9,7 +11,7 @@ import java.io.Serializable;
 public class Admin implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Integer id;
+    private Long id;
 
     private String name;
